requirements.txt
asgiref==3.2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-admin-honeypot==1.1.0
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.5.4
django-ckeditor==5.9.0
django-geojson==3.0.0
django-grappelli==2.14.1
django-import-export==2.0.2
django-js-asset==1.2.2
django-leaflet==0.26.0
et-xmlfile==1.0.1
Faker==4.0.3
jdcal==1.4.1
jsonfield==3.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
text-unidecode==1.3
xlrd==1.2.0
xlwt==1.3.0
